requirements.txt
torch>=2.0
transformers>=4.36.0
opencv-python
Pillow
